refactor: Header messages become Shell messages
Moving toward a single messages file for the entire Shell.
